+++ certification/internal/shell/has_required_labels.go
@@ -1,4 +1,4 @@
-package podmanexec
+package shell
 
 import (
 	"github.com/komish/preflight/certification"
