+++ certification/internal/shell/shell.go
@@ -1,4 +1,4 @@
-// Package podmanexec contains policy implementations that rely on utilizing
+// package shell contains policy implementations that rely on utilizing
 // Podman directly through the use of cmd.Exec. This implies that the
 // Podman CLI is installed and functional on a given system.
-package podmanexec
+package shell
